requirements.txt
matplotlib~=3.7.2
python-binance==1.0.19
cryptography==41.0.5
antlr4-python3-runtime==4.7.2
pika==1.3.2
websocket-client==1.7.0